requirements.txt
# Core
python-dotenv
flask

# LangChain + Google
langchain
langchain-google-genai
langchain-community

# Vector Store
chromadb




 